Added "support" for episodes and local files by assuming some default values for the values that aren't available in the JSON data. note support for episodes are just trash
data assumed:
- Songs
  - Icon
  - Album URI, for when caching the icon
- Episodes
  - Icon
  - Album URI, for when caching the icon
  - Duration, The length of the episode
  - Name

Added a iconNotFound.png as the assumed icon for albums and episodes that I cant get the cover of

changed name of iconCache to cacheFolder
made so the user can change the cacheFolder
started to add support for HoverOvert text for buttons
Now Storing an encrypted version of the SpotifyUserAccessToken

+++ src/main/java/me/zacharias/spotify/display/Time.java
@@ -49,7 +49,7 @@ public class Time {
     }
 
     public long getTimeInMilliseconds() {
-        return ms + (sec * 1000) + (min * 60 * 1000) + (hr * 60 * 60 * 1000);
+        return ms + (sec * 1000L) + ((long) min * 60 * 1000) + ((long) hr * 60 * 60 * 1000);
     }
 
     public int getTimeInMinutes() {
